rename marketplace to gridview and move marketplace specific stuff to marketplace component folder
